+++ src/quickPicks/quickPicks.ts
@@ -8,6 +8,7 @@ export * from './commitQuickPick';
 export * from './commitsQuickPick';
 export * from './commonQuickPicks';
 export * from './fileHistoryQuickPick';
+export * from './modesQuickPick';
 export * from './remotesQuickPick';
 export * from './repositoriesQuickPick';
 export * from './repoStatusQuickPick';