+++ src/statusBarController.ts
@@ -1,7 +1,7 @@
 'use strict';
 import { ConfigurationChangeEvent, Disposable, StatusBarAlignment, StatusBarItem, TextEditor, window } from 'vscode';
 import { Commands } from './commands';
-import { configuration, IConfig, StatusBarCommand } from './configuration';
+import { configuration, StatusBarCommand } from './configuration';
 import { isTextEditor } from './constants';
 import { Container } from './container';
 import { LinesChangeEvent } from './trackers/gitLineTracker';
@@ -9,8 +9,9 @@ import { CommitFormatter, GitCommit, ICommitFormatOptions } from './gitService';
 
 export class StatusBarController extends Disposable {
 
+    private _blameStatusBarItem: StatusBarItem | undefined;
     private _disposable: Disposable;
-    private _statusBarItem: StatusBarItem | undefined;
+    private _modeStatusBarItem: StatusBarItem | undefined;
 
     constructor() {
         super(() => this.dispose());
@@ -22,9 +23,10 @@ export class StatusBarController extends Disposable {
     }
 
     dispose() {
-        this.clear();
+        this.clearBlame();
 
-        this._statusBarItem && this._statusBarItem.dispose();
+        this._blameStatusBarItem && this._blameStatusBarItem.dispose();
+        this._modeStatusBarItem && this._modeStatusBarItem.dispose();
 
         Container.lineTracker.stop(this);
         this._disposable && this._disposable.dispose();
@@ -33,21 +35,48 @@ export class StatusBarController extends Disposable {
     private onConfigurationChanged(e: ConfigurationChangeEvent) {
         const initializing = configuration.initializing(e);
 
+        if (initializing || configuration.changed(e, configuration.name('mode').value)) {
+            const mode = Container.config.mode.active && Container.config.mode.statusBar.enabled
+                ? Container.config.modes[Container.config.mode.active]
+                : undefined;
+            if (mode && mode.statusBarItemName) {
+                const alignment = Container.config.mode.statusBar.alignment !== 'left' ? StatusBarAlignment.Right : StatusBarAlignment.Left;
+
+                if (configuration.changed(e, configuration.name('mode')('statusBar')('alignment').value)) {
+                    if (this._modeStatusBarItem !== undefined && this._modeStatusBarItem.alignment !== alignment) {
+                        this._modeStatusBarItem.dispose();
+                        this._modeStatusBarItem = undefined;
+                    }
+                }
+
+                this._modeStatusBarItem = this._modeStatusBarItem || window.createStatusBarItem(alignment, alignment === StatusBarAlignment.Right ? 999 : 1);
+                this._modeStatusBarItem.command = Commands.SwitchMode;
+                this._modeStatusBarItem.text = mode.statusBarItemName;
+                this._modeStatusBarItem.tooltip = `Switch GitLens Mode`;
+                this._modeStatusBarItem.show();
+            }
+            else {
+                if (this._modeStatusBarItem !== undefined) {
+                    this._modeStatusBarItem.dispose();
+                    this._modeStatusBarItem = undefined;
+                }
+            }
+        }
+
         if (!initializing && !configuration.changed(e, configuration.name('statusBar').value)) return;
 
-        const cfg = configuration.get<IConfig>();
-        if (cfg.statusBar.enabled) {
-            const alignment = cfg.statusBar.alignment !== 'left' ? StatusBarAlignment.Right : StatusBarAlignment.Left;
+        if (Container.config.statusBar.enabled) {
+            const alignment = Container.config.statusBar.alignment !== 'left' ? StatusBarAlignment.Right : StatusBarAlignment.Left;
 
             if (configuration.changed(e, configuration.name('statusBar')('alignment').value)) {
-                if (this._statusBarItem !== undefined && this._statusBarItem.alignment !== alignment) {
-                    this._statusBarItem.dispose();
-                    this._statusBarItem = undefined;
+                if (this._blameStatusBarItem !== undefined && this._blameStatusBarItem.alignment !== alignment) {
+                    this._blameStatusBarItem.dispose();
+                    this._blameStatusBarItem = undefined;
                 }
             }
 
-            this._statusBarItem = this._statusBarItem || window.createStatusBarItem(alignment, alignment === StatusBarAlignment.Right ? 1000 : 0);
-            this._statusBarItem.command = cfg.statusBar.command;
+            this._blameStatusBarItem = this._blameStatusBarItem || window.createStatusBarItem(alignment, alignment === StatusBarAlignment.Right ? 1000 : 0);
+            this._blameStatusBarItem.command = Container.config.statusBar.command;
 
             if (initializing || configuration.changed(e, configuration.name('statusBar')('enabled').value)) {
                 Container.lineTracker.start(
@@ -60,9 +89,9 @@ export class StatusBarController extends Disposable {
             if (configuration.changed(e, configuration.name('statusBar')('enabled').value)) {
                 Container.lineTracker.stop(this);
 
-                if (this._statusBarItem !== undefined) {
-                    this._statusBarItem.dispose();
-                    this._statusBarItem = undefined;
+                if (this._blameStatusBarItem !== undefined) {
+                    this._blameStatusBarItem.dispose();
+                    this._blameStatusBarItem = undefined;
                 }
             }
         }
@@ -74,7 +103,7 @@ export class StatusBarController extends Disposable {
         if (!e.pending && e.lines !== undefined) {
             const state = Container.lineTracker.getState(e.lines[0]);
             if (state !== undefined && state.commit !== undefined) {
-                this.updateStatusBar(state.commit, e.editor!);
+                this.updateBlame(state.commit, e.editor!);
 
                 return;
             }
@@ -83,54 +112,54 @@ export class StatusBarController extends Disposable {
         }
 
         if (clear) {
-            this.clear();
+            this.clearBlame();
         }
     }
 
-    async clear() {
-        if (this._statusBarItem !== undefined) {
-            this._statusBarItem.hide();
+    async clearBlame() {
+        if (this._blameStatusBarItem !== undefined) {
+            this._blameStatusBarItem.hide();
         }
     }
 
-    private updateStatusBar(commit: GitCommit, editor: TextEditor) {
+    private updateBlame(commit: GitCommit, editor: TextEditor) {
         const cfg = Container.config.statusBar;
-        if (!cfg.enabled || this._statusBarItem === undefined || !isTextEditor(editor)) return;
+        if (!cfg.enabled || this._blameStatusBarItem === undefined || !isTextEditor(editor)) return;
 
-        this._statusBarItem.text = `$(git-commit) ${CommitFormatter.fromTemplate(cfg.format, commit, {
+        this._blameStatusBarItem.text = `$(git-commit) ${CommitFormatter.fromTemplate(cfg.format, commit, {
             truncateMessageAtNewLine: true,
             dateFormat: cfg.dateFormat === null ? Container.config.defaultDateFormat : cfg.dateFormat
         } as ICommitFormatOptions)}`;
 
         switch (cfg.command) {
             case StatusBarCommand.ToggleFileBlame:
-                this._statusBarItem.tooltip = 'Toggle Blame Annotations';
+                this._blameStatusBarItem.tooltip = 'Toggle Blame Annotations';
                 break;
             case StatusBarCommand.DiffWithPrevious:
-                this._statusBarItem.command = Commands.DiffLineWithPrevious;
-                this._statusBarItem.tooltip = 'Compare Line Revision with Previous';
+                this._blameStatusBarItem.command = Commands.DiffLineWithPrevious;
+                this._blameStatusBarItem.tooltip = 'Compare Line Revision with Previous';
                 break;
             case StatusBarCommand.DiffWithWorking:
-                this._statusBarItem.command = Commands.DiffLineWithWorking;
-                this._statusBarItem.tooltip = 'Compare Line Revision with Working';
+                this._blameStatusBarItem.command = Commands.DiffLineWithWorking;
+                this._blameStatusBarItem.tooltip = 'Compare Line Revision with Working';
                 break;
             case StatusBarCommand.ToggleCodeLens:
-                this._statusBarItem.tooltip = 'Toggle Git CodeLens';
+                this._blameStatusBarItem.tooltip = 'Toggle Git CodeLens';
                 break;
             case StatusBarCommand.ShowQuickCommitDetails:
-                this._statusBarItem.tooltip = 'Show Commit Details';
+                this._blameStatusBarItem.tooltip = 'Show Commit Details';
                 break;
             case StatusBarCommand.ShowQuickCommitFileDetails:
-                this._statusBarItem.tooltip = 'Show Line Commit Details';
+                this._blameStatusBarItem.tooltip = 'Show Line Commit Details';
                 break;
             case StatusBarCommand.ShowQuickFileHistory:
-                this._statusBarItem.tooltip = 'Show File History';
+                this._blameStatusBarItem.tooltip = 'Show File History';
                 break;
             case StatusBarCommand.ShowQuickCurrentBranchHistory:
-                this._statusBarItem.tooltip = 'Show Branch History';
+                this._blameStatusBarItem.tooltip = 'Show Branch History';
                 break;
         }
 
-        this._statusBarItem.show();
+        this._blameStatusBarItem.show();
     }
 }